move InvokeResponse, WebRequest, WebResponse interfaces into separate files
