+++ libraries/botbuilder/src/channelServiceRoutes.ts
@@ -6,9 +6,11 @@
  * Licensed under the MIT License.
  */
 
-import { ChannelServiceHandler } from './channelServiceHandler';
 import { Activity, ConversationParameters, Transcript, AttachmentData } from 'botbuilder-core';
-import { WebRequest, WebResponse, StatusCodeError, StatusCodes } from './botFrameworkAdapter';
+
+import { ChannelServiceHandler } from './channelServiceHandler';
+import { StatusCodeError, StatusCodes } from './botFrameworkAdapter';
+import { WebRequest, WebResponse } from './interfaces';
 
 export type RouteHandler = (request: WebRequest, response: WebResponse) => void;
