+++ libraries/botbuilder/src/interfaces/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module botbuilder
+ */
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from './invokeResponse';
+export * from './webRequest';
+export * from './webResponse';
